backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Luc Elsby <luc@leverage-ai.co.uk>"]
readme = "README.md"
packages = [
    { include = "api", from = "src" },
    { include = "services", from = "src" },
    { include = "utils", from = "src" },
    { include = "tests", from = "src" }
]

[tool.poetry.dependencies]
python = "3.10"
fastapi = "0.115.4"
uvicorn = "0.29.0"
deepface = "0.0.93"
numpy = "1.26.4"
opencv-python = "4.10.0.84"
pillow = "10.2.0"
tensorboard = "2.9.1"
tensorboard-data-server = "0.6.1"
tensorboard-plugin-wit = "1.8.1"
tensorflow = "2.9.1"
tensorflow-estimator = "2.9.0"
tensorflow-io-gcs-filesystem = "0.31.0"
matplotlib = "3.8.1"
pandas = "2.2.3"
psutil = "6.1.0"
retina-face = "0.0.17"
watchdog = "6.0.0"
gql = "^3.5.2"
pytest = "^8.3.5"
pytest-asyncio = "^0.25.3"
requests-toolbelt = "^1.0.0"
aiohttp = "^3.11.14"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
